requirements.txt
absl-py==1.0.0
ale-py==0.8.0
ansiwrap==0.8.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.2
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
box2d-py==2.3.5
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
cloudpickle==2.0.0
colorama==0.4.4
coverage==6.3.2
cycler==0.11.0
Cython==0.29.33
dash==2.3.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
dm-tree==0.1.6
entrypoints==0.4
executing==0.8.3
fasteners==0.18
Flask==2.1.0
Flask-Compress==1.11
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.1.9
glfw==2.5.5
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.6.0
idna==3.3
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.11.3
importlib-resources==5.10.2
iniconfig==1.1.1
ipykernel==6.10.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.0
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
jupytext==1.13.7
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libclang==13.0.0
lz4==4.3.2
Markdown==3.3.6
markdown-it-py==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mdit-py-plugins==0.3.0
mistune==0.8.4
moviepy==1.0.3
mujoco==2.2.0
mujoco-py==2.1.2.14
mypy==0.942
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
papermill==2.3.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.6.0
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
Pygments==2.11.2
pyhocon==0.3.59
pylint==2.13.3
PyOpenGL==3.1.6
pyparsing==2.4.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
swig==4.1.1
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-gpu==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-probability==0.16.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
textwrap3==0.9.2
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.9.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typedload==2.15
typeguard==2.13.3
types-atomicwrites==1.4.3
types-requests==2.27.15
types-setuptools==57.4.11
types-tabulate==0.8.6
types-urllib3==1.26.11
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug @ https://github.com/pallets/werkzeug/archive/refs/tags/2.0.1.tar.gz
widgetsnbextension==3.6.0
wrapt==1.14.0
zipp==3.7.0
